pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zoltag"
version = "0.1.0"
description = "Multi-tenant image organization and search utility for Dropbox"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Zoltag Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Google Cloud Platform
    "google-cloud-storage>=2.14.0",
    "google-cloud-secret-manager>=2.18.0",
    "google-cloud-tasks>=2.15.0",
    "google-cloud-logging>=3.9.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    
    # Dropbox
    "dropbox>=11.36.0",
    
    # Image Processing
    "Pillow>=10.2.0",
    "pillow-heif>=0.15.0",
    "opencv-python-headless>=4.9.0",
    "piexif>=1.1.3",
    
    # ML/AI - Cost-optimized models
    "sentencepiece>=0.2.0",
    "sentence-transformers>=2.3.0",
    "imagehash>=4.3.1",
    
    # Search/Indexing
    "whoosh>=2.7.4",
    
    # Web Framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # Authentication
    "python-jose[cryptography]>=3.3.0",  # JWT verification via JWKS
    "slowapi>=0.1.9",  # Rate limiting for FastAPI

    # Utilities
    "pyyaml>=6.0.1",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "click>=8.1.0",
]

[project.scripts]
zoltag = "zoltag.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-PyYAML",
    "ipython>=8.20.0",
]

ml = [
    "face-recognition>=1.3.0",  # Requires cmake: brew install cmake
]
[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
zoltag = ["static/*", "static/**/*"]

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "--strict-markers",
    "--cov=zoltag",
    "--cov-report=term-missing",
]
asyncio_mode = "auto"
